+++ src/main/java/com/dasibom/practice/dto/DiarySaveReqDto.java
@@ -1,4 +1,4 @@
-package com.dasibom.practice.domain.dto;
+package com.dasibom.practice.dto;
 
 import com.dasibom.practice.domain.Pet;
 import com.dasibom.practice.domain.Stamp;
