+++ src/main/java/com/dasibom/practice/service/DiaryServiceImpl.java
@@ -5,8 +5,8 @@ import com.dasibom.practice.domain.DiaryStamp;
 import com.dasibom.practice.domain.Pet;
 import com.dasibom.practice.domain.Stamp;
 import com.dasibom.practice.domain.User;
-import com.dasibom.practice.domain.dto.DiaryDetailResDto;
-import com.dasibom.practice.domain.dto.DiarySaveReqDto;
+import com.dasibom.practice.dto.DiaryDetailResDto;
+import com.dasibom.practice.dto.DiarySaveReqDto;
 import com.dasibom.practice.repository.DiaryRepository;
 import com.dasibom.practice.repository.PetRepository;
 import com.dasibom.practice.repository.StampRepository;
